Fix linear color space conversion to use uint8 data type
Update the sRGB-to-linear color space conversion to return uint8 instead of float32, ensuring consistent data type handling throughout the optimization pipeline. This prevents potential type mismatches and improves memory efficiency.

+++ src/consumer/led_optimizer.py
@@ -209,8 +209,8 @@ class LEDOptimizer:
             np.power((srgb_normalized + 0.055) / 1.055, 2.4),
         )
 
-        # Scale back to [0, 255]
-        linear_frame = (linear_normalized * 255.0).astype(np.float32)
+        # Scale back to [0, 255] and convert to uint8
+        linear_frame = (linear_normalized * 255.0).astype(np.uint8)
 
         return linear_frame
